UI/requirements.txt
streamlit # streamlit>=1.37.01.31.0
pandas
requests>=2.31.0
google-auth-oauthlib>=1.2.0
firebase-admin>=7.1.0
streamlit-js-eval # extra-streamlit-components
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-dotenv==1.0.1
# google-cloud-storage>=2.18.0
# watchdog>=6.0.0
jinja2==3.1.4
python-multipart==0.0.9
authlib>=1.6.1
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0